+++ src/menu/ui_components/background.c
